requirements.txt
configparser==5.3.0
discord.py==2.1.0
SQLAlchemy==1.4.42
PyMySQL==1.0.2
cffi==1.15.1
PyYAML==6.0